Create allennlp directory
